+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/TigerCucumberRunner.java
@@ -4,7 +4,6 @@
 
 package de.gematik.test.tiger;
 
-import de.gematik.test.tiger.lib.TigerDirector;
 import io.cucumber.core.options.CommandlineOptionsParser;
 import io.cucumber.core.options.RuntimeOptions;
 import io.cucumber.core.plugin.SerenityReporter;
@@ -23,27 +22,16 @@ import org.junit.runners.model.InitializationError;
 /**
  * When started via Intellij the main method is run.
  * When started via maven the constructor gets called for each driver class.
- * Later must have the TigerCucumberListener registered as plugin in the cucumberoptions
- * When using the tiger maven plugin this is taken care of automgically!
- *
- * First one sets it in the code explicitely.
+ * Later must have the TigerCucumberListener registered as plugin in the cucumber-options
+ * When using the tiger maven plugin this is taken care of automagically!
+ * First one sets it in the code explicitly.
  */
+
 @Slf4j
 public class TigerCucumberRunner extends CucumberSerenityRunner {
 
-    private static RuntimeException tigerStartupFailedException;
-
     public static void main(String[] argv) {
         log.info("Starting TigerCucumberRunner.main()...");
-        if (tigerStartupFailedException != null) {
-            log.error("Aborting due to earlier errors in setting up Tiger!");
-            System.exit(1);
-        }
-        initializeTiger();
-        if (tigerStartupFailedException != null) {
-            log.error("Unable to start Tiger!", tigerStartupFailedException);
-            System.exit(1);
-        }
         Supplier<ClassLoader> classLoaderSupplier = ClassLoaders::getDefaultClassLoader;
         byte exitstatus = run(argv, classLoaderSupplier);
         System.exit(exitstatus);
@@ -67,22 +55,6 @@ public class TigerCucumberRunner extends CucumberSerenityRunner {
     public TigerCucumberRunner(Class clazz) throws InitializationError {
         super(clazz);
         log.info("Starting TigerCucumberRunner for {}", clazz.getName());
-        if (tigerStartupFailedException != null) {
-            throw new InitializationError(tigerStartupFailedException);
-        }
-        initializeTiger();
-        if (tigerStartupFailedException != null) {
-            throw new InitializationError(tigerStartupFailedException);
-        }
-    }
-
-    private synchronized static void initializeTiger() {
-        try {
-            TigerDirector.registerShutdownHook();
-            TigerDirector.start();
-        } catch (RuntimeException rte) {
-            tigerStartupFailedException = rte;
-        }
     }
 
     public static Runtime using(Supplier<ClassLoader> classLoaderSupplier, RuntimeOptions runtimeOptions) {